check var value

+++ gitops-lab/teardown-lab-environment.sh
@@ -5,6 +5,11 @@ if [[ -z "$TF_VAR_namespace" ]]; then
     exit 1
 fi
 
+if [[ "$TF_VAR_namespace" == *"<"* ]]; then
+    echo "Please don't paste the example, replace with your namespace" 1>&2
+    exit 1
+fi
+
 cd ~/terraform-modular-demo-framework
 
 terragrunt run-all destroy --terragrunt-modules-that-include ./gitops-lab.hcl --terragrunt-non-interactive
